src/third_party/blink/web_tests/http/tests/contacts/idl-NavigatorContacts.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_true('contacts' in navigator,
              'navigator.contacts exists in navigator.');
  assert_true('ContactInfo' in window,
              'ContactInfo exists in window.');
}, 'navigator.contacts IDL test');
</script>
